+++ search-box@mtwebster/files/search-box@mtwebster/CHANGELOG.md
@@ -1,3 +1,11 @@
+### v2.2.0~20240615
+  * New settings layout.
+  * Fixed a bug that prevented you from modifying the list of search engines.
+  * Removed deprecated versions.
+
+### v2.1.0~20240614
+  * You can now select the background color and text color of the search box.
+
 ### v2.0.1~20240520
   * From now on, you can select the search engines you want to show in the context menu.
   * The current search engine is preceded by a dot in the context menu.